Refactor date creation
